requirements.txt
numpy==1.24.3
pandas==2.0.3
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
Pillow==10.1.0
plotly==5.20.0
scikit-learn==1.3.2
scipy==1.10.1
torch==2.1.2
torchvision==0.16.2
